+++ packages/typescript/CHANGELOG.md
@@ -10,9 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update `jsdoc/require-jsdoc` to loosen requirements for various kinds of symbols ([#433](https://github.com/MetaMask/eslint-config/pull/433))
-  - Instead of requiring JSDoc for all arrow functions and function expressions, require only arrow functions not contained within plain objects or are not arguments to functions or methods.
-  - Instead of requiring JSDoc for all interfaces or type aliases, require only those that do not appear in declare blocks (ambient declarations).
-  - Instead of requiring JSDoc for all object types in return types, require only object types in "root" types.
+  - Require JSDoc not for all arrow functions, but only those not contained within plain objects or are not arguments to functions or methods.
+  - Require JSDoc not for all interfaces or type aliases, but only those that do not appear in declare blocks (ambient declarations).
+  - Require JSDoc not for all object types in return types, but only those in "root" types.
 
 ## [15.0.0]