+++ packages/typescript/CHANGELOG.md
@@ -10,9 +10,62 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update `jsdoc/require-jsdoc` to loosen requirements for various kinds of symbols ([#433](https://github.com/MetaMask/eslint-config/pull/433))
-  - Require JSDoc not for all arrow functions, but only those not contained within plain objects or are not arguments to functions or methods.
-  - Require JSDoc not for all interfaces or type aliases, but only those that do not appear in declare blocks (ambient declarations).
-  - Require JSDoc not for all object types in return types, but only those in "root" types.
+  - JSDoc is no longer required for arrow functions or function expressions which are values of object properties:
+    ```typescript
+    const foo = {
+      // This arrow function is no longer required to be documented
+      bar: () => {
+        // ...
+      }
+    });
+
+    const foo = {
+      // This function expression is no longer required to be documented
+      bar: function() {
+        // ...
+      }
+    });
+    ```
+  - JSDoc is no longer required for arrow functions or function expressions which are arguments to functions or methods:
+    ```typescript
+    // This arrow function is no longer required to be documented
+    foo(() => {
+      // ...
+    })
+
+    // This function expression is no longer required to be documented
+    foo(function () {
+      // ...
+    })
+    ```
+  - JSDoc is no longer required for interfaces, type aliases, or enums that appear in `declare` blocks (even if they are exported):
+    ```typescript
+    declare module 'some-module' {
+      // This type is no longer required to be documented
+      type Bar = 'baz';
+
+      // This interface is no longer required to be documented, even though it
+      // is exported
+      export interface Y {
+        // whatever
+      }
+
+      // This enum is no longer required to be documented
+      enum Fooz {
+        Bar = 'baz',
+      }
+    }
+    ```
+  - JSDoc is no longer required for inline object types in return types:
+    ```typescript
+    // This object type is no longer required to be documented
+    function foo(): {
+      bar: 'baz';
+      qux: 'blargh';
+    } {
+      // ...
+    }
+    ```
 
 ## [15.0.0]